builds with Dynamixel2Arduino + OpenCM
The update converts the code to use the Robotis Dynamixel Library
Dynamixel2Arduino.  You can install this library using the
Arduino library manager or from github.
https://github.com/ROBOTIS-GIT/Dynamixel2Arduino

This library is discussed as part of the Dynamixel Shield emanual:
https://emanual.robotis.com/docs/en/parts/interface/dynamixel_shield/#library-api

This is just step 1, probably a lot more refinements wil be added.
But as part of this it removes the usage of my older library:
BioloidSerial.  We are still using the Pose Interpolation stuff, which
I have merged into this sketch.

With this change, it makes it easer to support this sketch on some
different platforms such as the OpenCM9.04, which I have started
working.  Also this library is built to support both Protocol 1 and
Protocol 2 libraries.

So with these changes I am now experimenting on a Beta version of the
Trossen Robotics PhantomX IV Robot with XL430-W250 servos. They are
now talking... Still lots more debugging to go.