Bioloid - When turn servos on detect HW errors and reboot servos
We were having issues with the servos going all the way up, and then if we logically turn servos off and then back on, they don't recover.

What happened was servo got into an error state and then refused to move.  So as part of the Make sure servos are on, it look at the HW status bytes and if not zero it resets that servo...